eshop_server/user_message/serializers.py
from rest_framework import serializers

from user_message.models import UserMessage


class UserMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserMessage
        fields = '__all__'
